Move Queue & Command to CommandQueue
